requirements.txt
django==2.1.1
django-cors-headers==3.2.0
djangorestframework==3.10.3
requests==2.22.0
googlemaps==4.0.0
mysqlclient==1.3.12
django-heroku
gunicorn
